Separar processamento
